+++ tests/testthat/test-empty.R
@@ -1,7 +1,0 @@
-test_that("nothing tested yet", {
-  withr::local_file(file.path(tempdir(), "test1"))
-
-  ## TODO: need tests
-
-  withr::deferred_run()
-})